+++ libs/admin/test/wagoe/admin/shell/http_test.clj
@@ -17,7 +17,7 @@
             [wagoe.admin.shell.service :as service]
             [wagoe.admin.shell.schema-repository :as schema-repo]
             [wagoe.platform.shell.adapters.database.factory :as db-factory]
-            [wagoe.platform.shell.adapters.database.common.execution :as db]
+            [wagoe.platform.database :as db]
             [wagoe.observability.logging.shell.adapters.no-op :as logging-no-op]
             [wagoe.observability.errors.shell.adapters.no-op :as error-reporting-no-op]
             [wagoe.shared.ui.core.components :as ui-components]